+++ packages/example-react/src/router.tsx
@@ -1,6 +1,7 @@
 import { createBrowserRouter } from 'react-router-dom'
 import { App } from './App'
 import { Example } from './Example'
+import { ExampleLegacy } from './ExampleLegacy'
 import { Login } from './Login'
 import { Logout } from './Logout'
 
@@ -10,6 +11,7 @@ export const router = createBrowserRouter([
     element: <App />,
     children: [
       { path: 'example', element: <Example /> },
+      { path: 'example-legacy', element: <ExampleLegacy /> },
       { path: 'login', element: <Login /> },
       { path: 'logout', element: <Logout /> },
     ],